move trace attribute serializer to tyr
