+++ Makefile.am
@@ -383,6 +383,7 @@ libvalhalla_la_SOURCES = \
 	src/tyr/height_serializer.cc \
 	src/tyr/locate_serializer.cc \
 	src/tyr/transit_available_serializer.cc \
+	src/tyr/trace_serializer.cc \
 	src/tyr/navigator.cc \
 	src/tyr/actor.cc
 libvalhalla_la_CPPFLAGS = @BOOST_CPPFLAGS@ @RAPIDJSON_CPPFLAGS@ $(DEPS_CFLAGS) $(SERVICE_DEPS_CFLAGS)
