+++ src/thor/trace_route_action.cc
@@ -35,6 +35,12 @@ struct MapMatch {
   int edge_index = -1;
 };
 
+// <Confidence score, raw score, match results, trip path> tuple indexes
+constexpr size_t kConfidenceScoreIndex = 0;
+constexpr size_t kRawScoreIndex = 1;
+constexpr size_t kMatchResultsIndex = 2;
+constexpr size_t kTripPathIndex = 3;
+
 }
 
 namespace valhalla {
